+++ Expo1900/Expo1900/KoreaWorksViewController.swift
@@ -7,9 +7,53 @@
 
 import UIKit
 
-class KoreaWorksViewController: UIViewController, UITableViewDelegate, UITableViewDataSource {
+class KoreaWorksViewController: UIViewController {
     
     private var koreaWorksList: [KoreaWorks] = []
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        decodeKoreaWorks()
+        self.navigationController?.setNavigationBarHidden(false, animated: true)
+    }
+    
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        guard segue.identifier == "showDetail" else {
+            return
+        }
+        guard let destinationController = segue.destination as? KoreaWorksDetailViewController,
+              let index = sender as? Int else {
+            return
+        }
+        let koreaWorksInfo = koreaWorksList[index]
+        destinationController.koreaWorksInfo = koreaWorksInfo
+    }
+}
+
+extension KoreaWorksViewController {
+    
+    private func decodeKoreaWorks() {
+        
+        let jsonDecoder = JSONDecoder()
+        guard let assetData: NSDataAsset = NSDataAsset(name: "items") else {
+            return
+        }
+        guard let koreaWorksData = try? jsonDecoder.decode([KoreaWorks].self, from: assetData.data) else {
+            return
+        }
+        koreaWorksList = koreaWorksData
+    }
+}
+
+extension KoreaWorksViewController: UITableViewDelegate {
+    
+    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        performSegue(withIdentifier: "showDetail", sender: indexPath.row)
+        tableView.deselectRow(at: indexPath, animated: false)
+    }
+}
+
+extension KoreaWorksViewController: UITableViewDataSource {
     
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         return koreaWorksList.count
@@ -21,56 +65,7 @@ class KoreaWorksViewController: UIViewController, UITableViewDelegate, UITableVi
             return UITableViewCell()
         }
         let koreaWorksInfo = koreaWorksList[indexPath.row]
-        cell.updatdCell(info: koreaWorksInfo)
+        cell.updateCell(info: koreaWorksInfo)
         return cell
     }
-    
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        performSegue(withIdentifier: "showDetail", sender: indexPath.row)
-        tableView.deselectRow(at: indexPath, animated: false)
-    }
-    
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        if segue.identifier == "showDetail" {
-            let a = segue.destination as? KoreaWorksDetailViewController
-            if let index = sender as? Int {
-                let koreaWorksInfo = koreaWorksList[index]
-                a?.koreaWorksInfo = koreaWorksInfo
-            }
-        }
-    }
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        decodeKoreaWorks()
-        self.navigationController?.setNavigationBarHidden(false, animated: true)
-    }
 }
-
-class ListCell: UITableViewCell {
-    @IBOutlet weak var imgView: UIImageView!
-    @IBOutlet weak var nameLabel: UILabel!
-    @IBOutlet weak var shortDescriptionLabel: UILabel!
-    
-    func updatdCell(info: KoreaWorks) {
-        imgView.image = info.image
-        nameLabel.text = info.name
-        shortDescriptionLabel.text = info.shortDescription
-    }
-}
-
-extension KoreaWorksViewController {
-    
-    private func decodeKoreaWorks() {
-        
-        let jasonDecoder = JSONDecoder()
-        guard let assetData: NSDataAsset = NSDataAsset(name: "items") else {
-            return
-        }
-        guard let koreaWorksData = try? jasonDecoder.decode([KoreaWorks].self, from: assetData.data) else {
-            return
-        }
-        koreaWorksList = koreaWorksData
-    }
-}
-
